% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{add_two}
\alias{add_two}
\title{Addition}
\usage{
add_two(a, b)
}
\arguments{
\item{Erster}{Summand a}

\item{Zweiter}{Summand b}
}
\value{
Summe beider Summanden
}
\description{
Addition
}
\examples{
add_two(4, 3)
add_two(-5, 3)
}
