% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplikation.R
\name{multiply_two}
\alias{multiply_two}
\title{Multiplication}
\usage{
multiply_two(a, b)
}
\arguments{
\item{Multiplikator}{(numerisch)}

\item{Multiplikand}{(numerisch)}
}
\value{
Produkt der beiden Faktoren
}
\description{
Multiplication
}
\examples{
multiply_two(2, 3)
multiply_two(4, -2)

}
