% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtract_two}
\alias{subtract_two}
\title{Substraction}
\usage{
subtract_two(a, b)
}
\arguments{
\item{Minuend}{(numerisch)}

\item{Subtrahend}{(numerisch)}
}
\value{
Die Differenz der beiden Eingangswerte.
}
\description{
Substraction
}
\examples{
subtract_two(3, 1)
subtract_two(-1, 3)

}
